[Kernel] Phase 4 (layernorm): move JIT layernorm operators into kernels.ops.layernorm (RFC #29630)
Moves fused_eh_norm rmsnorm_hf from sglang.jit_kernel into sglang.kernels.ops.layernorm (classified);
compat shims at old paths (removed Phase 5); csrc stays in jit_kernel.
